+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - Unreleased
+
+### Added
+
+- Support for `"midnight"` and `"midday"` keywords in `parseTime()` and `parseDateTime()`.
+- Support for `"yesterday"` keyword in `parseDate()` and `parseDateTime()`.
+
 ## [2.1.0] - 2026-01-26
 
 ### Added